ref: shared features data entity
