+++ api/apps/api/src/modules/geo-features/geo-feature.geo.entity.ts
@@ -1,9 +1,7 @@
-import { ApiProperty } from '@nestjs/swagger';
-import { Geometry } from 'geojson';
-import { Column, Entity, PrimaryColumn, ViewEntity } from 'typeorm';
 import { BaseServiceResource } from '@marxan-api/types/resource.interface';
-
-import { GeometrySource } from './geometry-source.enum';
+import { GeoFeatureGeometry } from '@marxan/geofeatures';
+import { ApiProperty } from '@nestjs/swagger';
+import { Column, ViewEntity } from 'typeorm';
 
 export const geoFeatureResource: BaseServiceResource = {
   className: 'GeoFeature',
@@ -13,29 +11,6 @@ export const geoFeatureResource: BaseServiceResource = {
   },
   moduleControllerPrefix: 'geo-features',
 };
-
-@Entity('features_data')
-export class GeoFeatureGeometry {
-  @ApiProperty()
-  @PrimaryColumn()
-  id!: string;
-
-  @Column('geometry', { name: 'the_geom' })
-  theGeom?: Geometry;
-
-  @Column('jsonb')
-  properties?: Record<string, string | number>;
-
-  @Column(`enum`, {
-    enum: GeometrySource,
-    nullable: true,
-  })
-  source?: GeometrySource | null;
-
-  @ApiProperty()
-  @Column('uuid', { name: 'feature_id' })
-  featureId?: string;
-}
 
 export class JSONAPIGeoFeaturesGeometryData {
   @ApiProperty()
