+++ api/apps/api/src/modules/geo-features/geo-features.module.ts
@@ -2,10 +2,8 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Project } from '@marxan-api/modules/projects/project.api.entity';
 import { GeoFeature } from './geo-feature.api.entity';
-import {
-  GeoFeatureGeometry,
-  GeoFeaturePropertySet,
-} from './geo-feature.geo.entity';
+import { GeoFeatureGeometry } from '@marxan/geofeatures';
+import { GeoFeaturePropertySet } from './geo-feature.geo.entity';
 
 import { GeoFeaturesController } from './geo-features.controller';
 import { GeoFeaturesService } from './geo-features.service';
