+++ api/apps/api/test/jest-e2e.json
@@ -68,6 +68,8 @@
     "@marxan/cloning/(.*)": "<rootDir>/../../../libs/cloning/src/$1",
     "@marxan/cloning": "<rootDir>/../../../libs/cloning/src",
     "@marxan/webshot/(.*)": "<rootDir>../../../libs/webshot/src/$1",
-    "@marxan/webshot": "<rootDir>../../../libs/webshot/src"
+    "@marxan/webshot": "<rootDir>../../../libs/webshot/src",
+    "@marxan/geofeatures/(.*)": "<rootDir>../../../libs/geofeatures/src/$1",
+    "@marxan/geofeatures": "<rootDir>../../../libs/geofeatures/src"
   }
 }
