+++ api/apps/geoprocessing/src/modules/features/features.module.ts
@@ -4,7 +4,7 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { FeaturesController } from './features.controller';
 import { FeatureService } from './features.service';
 import { TileModule } from '@marxan-geoprocessing/modules/tile/tile.module';
-import { GeoFeatureGeometry } from '@marxan-geoprocessing/modules/features/features.geo.entity';
+import { GeoFeatureGeometry } from '@marxan/geofeatures';
 import { WorkerModule } from '@marxan-geoprocessing/modules/worker';
 import {
   copyQueueNameProvider,
