+++ api/apps/geoprocessing/test/jest-e2e.json
@@ -70,6 +70,8 @@
     "@marxan/files-repository/(.*)": "<rootDir>/../../libs/files-repository/src/$1",
     "@marxan/files-repository": "<rootDir>/../../libs/files-repository/src",
     "@marxan/cloning/(.*)": "<rootDir>/../../libs/cloning/src/$1",
-    "@marxan/cloning": "<rootDir>/../../libs/cloning/src"
+    "@marxan/cloning": "<rootDir>/../../libs/cloning/src",
+    "@marxan/geofeatures/(.*)": "<rootDir>../../../libs/geofeatures/src/$1",
+    "@marxan/geofeatures": "<rootDir>../../../libs/geofeatures/src"
   }
 }
