+++ api/libs/geofeatures/src/index.ts
@@ -1,0 +1,2 @@
+export { GeoFeatureGeometry } from './geo-feature.geo.entity';
+export { GeometrySource } from './geometry-source.enum';
