+++ api/tsconfig.json
@@ -174,6 +174,12 @@
       ],
       "@marxan/webshot/*": [
         "libs/webshot/src/*"
+      ],
+			"@marxan/geofeatures": [
+				"libs/geofeatures/src"
+			],
+			"@marxan/geofeatures/*": [
+        "libs/geofeatures/src/*"
       ]
     }
   },
